backend/requirements.txt
python-dotenv
pyodbc
asyncua
fastapi
uvicorn[standard]
pydantic
pydantic-core
typing-extensions